Move entclient package methods
Split ent client package methods into a separate `package` file.
